2610: Make a separate component for ExportEventButton
